tests/requirements.txt
robotframework >= 7.1
pyrenode3 @ git+https://github.com/antmicro/pyrenode3.git@3892655
